[turbopack] Seperate `JsValue` into it's own folder in the analyzer
