+++ turbopack/crates/turbopack-ecmascript/src/analyzer/mod.rs
@@ -441,7 +441,6 @@ mod tests {
     use std::{mem::take, path::PathBuf, time::Instant};
 
     use parking_lot::Mutex;
-    use rstest::rstest;
     use rustc_hash::FxHashMap;
     use swc_core::{
         common::{Mark, comments::SingleThreadedComments},
@@ -930,205 +929,5 @@ mod tests {
         })
         .await
         .unwrap()
-    }
-
-    #[test]
-    #[cfg(target_pointer_width = "64")]
-    fn jsvalue_size() {
-        assert_eq!(32, size_of::<JsValue>());
-    }
-
-    #[test]
-    fn is_string_constant() {
-        let value = EvalContext::eval_single_expr_lit(&rcstr!("'hello'")).unwrap();
-        assert_eq!(value.is_string(), Some(true));
-    }
-
-    #[rstest]
-    #[case("1 && 'hello'")]
-    #[case("'hello' || 'bye' || 2")]
-    fn is_string_short_circuiting_positive(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_string(),
-            Some(true),
-            "expected '{}' to be a string",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("'hello' && 2")]
-    #[case("2 || 1 || 'hello' || 'bye'")]
-    fn is_string_short_circuiting_negative(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_string(),
-            Some(false),
-            "expected '{}' not to be a string",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("x && 2")]
-    #[case("1 && x")]
-    #[case("1 && 'a' && x")]
-    #[case("x || 'bye'")]
-    #[case("false || x")]
-    fn is_string_short_circuiting_unknown(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_string(),
-            None,
-            "expected to be unable to determine whether '{}' is a string",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("'' && 'string'")]
-    #[case("false || ''")]
-    #[case("1 && 'a' && ''")]
-    fn is_empty_string_short_circuiting_positive(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_empty_string(),
-            Some(true),
-            "expected '{}' to be an empty string",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("false && ''")]
-    #[case("'' || 'string'")]
-    #[case("'' || 0 || 'string'")]
-    fn is_empty_string_short_circuiting_negative(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_empty_string(),
-            Some(false),
-            "expected '{}' not to be an empty string",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("x && ''")]
-    #[case("1 && x")]
-    #[case("x || ''")]
-    #[case("'' || x")]
-    #[case("false || 0 || x")]
-    fn is_empty_string_short_circuiting_unknown(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_empty_string(),
-            None,
-            "expected to be unable to determine whether '{}' is an empty string",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("null && ''")]
-    #[case("'' || null")]
-    #[case("1 && 2 && null")]
-    fn is_nullish_short_circuiting_positive(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_nullish(),
-            Some(true),
-            "expected '{}' to be nullish",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("'' && null")]
-    #[case("null || ''")]
-    #[case("null || '' || 'a'")]
-    fn is_nullish_short_circuiting_negative(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_nullish(),
-            Some(false),
-            "expected '{}' not to be nullish",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("x && null")]
-    #[case("1 && x")]
-    #[case("x || null")]
-    #[case("null || x")]
-    #[case("false || x")]
-    #[case("1 && x && null")]
-    fn is_nullish_short_circuiting_unknown(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_nullish(),
-            None,
-            "expected to be unable to determine whether '{}' is nullish",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("'' && null")]
-    #[case("null || ''")]
-    #[case("null || 0 || 'a'")]
-    fn is_not_nullish_short_circuiting_positive(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_not_nullish(),
-            Some(true),
-            "expected '{}' to be not-nullish",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("null && ''")]
-    #[case("'' || null")]
-    #[case("'' || 0 || null")]
-    fn is_not_nullish_short_circuiting_negative(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_not_nullish(),
-            Some(false),
-            "expected '{}' not to be not-nullish",
-            input
-        );
-    }
-
-    #[rstest]
-    #[case("x && null")]
-    #[case("1 && x")]
-    #[case("x || null")]
-    #[case("null || x")]
-    #[case("false || x")]
-    #[case("false || x || ''")]
-    fn is_not_nullish_short_circuiting_unknown(#[case] input: &str) {
-        assert_eq!(
-            EvalContext::eval_single_expr_lit(&input.into())
-                .unwrap()
-                .is_not_nullish(),
-            None,
-            "expected to be unable to determine whether '{}' is not-nullish",
-            input
-        );
     }
 }
